backend/requirements.txt
asgiref==3.3.1
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3
Django==3.1.3
django-cors-headers==3.5.0
django-filter==2.4.0
django-rest-swagger==2.2.0
django-simple-history==2.12.0
django-utils==0.0.2
django-utils-six==2.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangoutils==0.1.1
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
openapi-codec==1.3.2
pycparser==2.20
PyJWT==1.7.1
PyMySQL==0.10.1
pytz==2020.4
requests==2.25.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.2
uvicorn==0.13.4
